Switch ruleNames to use std::vector<std::wstring>.
Changes the list of rule and token names to use a vector instead of an
array pointer.  This greatly simplifies memory management throughout
the code.  This particular design assumes we’ll be using C++11.  If
this is not the case, we can make the moves explicit (using std::move)
but that seems like a good thing to do if there is a future need to
support older/simpler compilers.